third_party/WebKit/LayoutTests/html/dialog/non-modal-dialog-layout-expected.txt

Tests layout of non-modal dialogs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Test absolute position
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top + offset
PASS dialog.getBoundingClientRect().left is relativeContainer.getBoundingClientRect().left + offset

Test static position
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top
PASS dialog.getBoundingClientRect().left is relativeContainer.getBoundingClientRect().left

Test relative position
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top + offset
PASS dialog.getBoundingClientRect().left is relativeContainer.getBoundingClientRect().left + offset

Test fixed position
PASS dialog.getBoundingClientRect().top is offset
PASS dialog.getBoundingClientRect().left is offset
PASS successfullyParsed is true

TEST COMPLETE
